api/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
argcomplete==3.6.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.7.0
eventlet==0.39.0
exceptiongroup==1.3.1
fastapi==0.104.1
Flask==2.3.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
gotrue==2.12.4
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
http-ece==1.2.1
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==11.1.0
postgrest==0.18.0
psycopg2-binary==2.9.10
py-vapid==1.9.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pystray==0.19.5
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.3
python-escpos==3.1
python-socketio==5.15.0
pywebpush==1.14.1
PyYAML==6.0.3
qrcode==8.0
realtime==2.25.1
requests==2.31.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.24.0
waitress==3.0.2
websockets==15.0.1
Werkzeug==2.3.0
wsproto==1.3.2